+++ parametric_scripts/sparkle.py
@@ -16,7 +16,7 @@ from backend.util import color_range, dim_color
 
 def run(**kwargs):
     color_input = kwargs["color"]
-    interval = int(kwargs["interval"])
+    interval = float(kwargs["interval"])
 
     ceil = Ceiling(auto_write=False)
     ceil.use_linear()
