+++ parametric_scripts/spinning.py
@@ -11,7 +11,7 @@ from backend.util import color_range, dim_color
 
 def run(**kwargs):
     color_input = kwargs["color"]
-    interval = int(kwargs["interval"])
+    interval = float(kwargs["interval"])
 
     ceil = Ceiling()
     ceil.use_polar((0.5, 0.5))
